requirements.txt
blinker==1.6.2
click==8.1.3
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
marshmallow==3.19.0
packaging==23.1
phonenumbers==8.13.14
pluggy==1.0.0
psycopg2==2.9.6
pytest==7.3.2
python-dateutil==2.8.2
six==1.16.0
SQLAlchemy==2.0.16
typing_extensions==4.6.3
Werkzeug==2.3.6
